work on tasks

+++ src/projects/projects.service.ts
@@ -2,6 +2,7 @@ import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateProjectDto } from './dto/create-project.dto';
 import { UsersService } from 'src/users/users.service';
+import { UpdateProjectDto } from './dto/update-project.dto';
 
 @Injectable()
 export class ProjectsService {
@@ -46,11 +47,89 @@ export class ProjectsService {
   }
 
   async deleteProject(tenantId: string, projectId: string) {
-    return await this.prisma.projects.delete({
+    const isProjectExist = await this.getProjectById(tenantId, projectId);
+    if (!isProjectExist) {
+      throw new HttpException(
+        {
+          success: false,
+          message: 'Project not exist',
+        },
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+    const project = await this.prisma.projects.delete({
       where: {
         id: BigInt(projectId),
         tenantId: BigInt(tenantId),
       },
     });
+    return {
+      ...project,
+      id: project.id.toString(),
+      tenantId: project.tenantId.toString(),
+    };
+  }
+
+  async getProjectById(tenantId: string, projectId: string) {
+    try {
+      const project = await this.prisma.projects.findFirstOrThrow({
+        where: { id: BigInt(projectId), tenantId: BigInt(tenantId) },
+      });
+      return {
+        ...project,
+        id: String(project.id),
+        tenantId: String(project.tenantId),
+      };
+    } catch (err) {
+      throw new HttpException(
+        {
+          success: false,
+          message: 'Project not found',
+          status: HttpStatus.NOT_FOUND,
+        },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+  }
+
+  async updateProjectDetails(
+    tenantId: string,
+    projectId: string,
+    projectDetails: UpdateProjectDto,
+  ) {
+    const isProjectExist = await this.getProjectById(tenantId, projectId);
+    if (!isProjectExist) {
+      throw new HttpException(
+        {
+          success: false,
+          message: 'Project not found',
+          status: HttpStatus.NOT_FOUND,
+        },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    const { name, description } = projectDetails;
+    const newDetails: UpdateProjectDto = {};
+    if (name) {
+      newDetails.name = name;
+    }
+    if (description) {
+      newDetails.description = description;
+    }
+    const project = await this.prisma.projects.update({
+      where: {
+        id: BigInt(projectId),
+        tenantId: BigInt(tenantId),
+      },
+      data: {
+        ...newDetails,
+      },
+    });
+    return {
+      ...project,
+      id: String(project.id),
+      tenantId: String(project.tenantId),
+    };
   }
 }
